LayoutTests/imported/w3c/web-platform-tests/FileAPI/url/cross-global-revoke.sub-expected.txt


PASS It is possible to revoke same-origin blob URLs from different frames.
PASS It is possible to revoke same-origin blob URLs from a different worker global.
PASS It is not possible to revoke cross-origin blob URLs.
